chore(#3096): move the WS0 measurement rig to #3272; keep the CI fixture self-contained
Owner-ordered SPLIT (2026-08-03). A sliced roborev pass left 7 blockers open at
HEAD, 5 of them in `scripts/perf/`, three being rig guards that earlier fix rounds
made fail-open or bypassable. The rig needs its own review footing and must not
hold the reviewed core, so it is re-anchored to issue #3272.

REMOVED here (moves to #3272):

* `tools/ws0-corpus-gen/` — the whole crate (generator CLI, bare-scan bench,
  corpus-identity record). It was a workspace member only via the root
  `Cargo.toml`'s `tools/*` glob, so deleting the directory de-registers it; the
  root manifest is otherwise back to `origin/main` (its `sha2` workspace entry
  existed solely so the generator and cqlite-core could not drift, and
  cqlite-core returns to its own `sha2 = "0.10"`).
* `scripts/perf/` — driver, CPU-pinning library, reporter, README.
* `scripts/tests/test_ws0_report_guards.sh` AND the `agent-gate.sh` `tooling-tests`
  hook that ran it; `scripts/agent-gate.sh` is restored to `origin/main` verbatim,
  so the gate references nothing that was removed.
* The `ws0-corpus-gen` dev-dependency of `cqlite-flight`, and its `Cargo.lock`
  entries (lock REGENERATED via cargo, not hand-edited — it is now byte-identical
  to `origin/main`).

KEPT, and still self-contained: the two CI tests that used the generator as their
fixture builder — `tests/issue_3096_arrow_buffer_digest.rs` (the Arrow-buffer
digest oracle) and `tests/issue_3096_framing_subphase.rs` (the IPC-framing
attribution test). Their fixture now comes from the new
`cqlite-flight/tests/support/ws0_fixture.rs` support module, which carries the
MINIMUM the fixture needs — pinned `ws0.events` DDL/schema, the SplitMix64 row
synthesis, the token-ordered `SSTableWriter` write — with the row content, schema
and write path UNCHANGED from the form the pinned digest was recorded against.
Measured: the CI-fixture digest is still 0xd0014e42e893f87f over 500 rows in 4
batches, 12 columns, 12.0 cells/row, identical on the merge and bypass arms. The
crate's DDL-pin unit test is retained as `assert_ddl_matches_the_committed_pin()`,
called by the oracle, so a drifted DDL still fails.

The oracle's OPT-IN measurement-corpus arm needs only an env var plus a corpus
root, so it still works, but the generator that PRODUCES such a corpus lives in
#3272 — its unset-env skip message now says so explicitly. SET-but-unusable
remains a hard failure (never a silent skip).
